Check structure of data in downloaded feedback CSV
