+++ cypress/e2e/common/admin.js
@@ -1,3 +1,4 @@
+const path = require('path');
 import { Then } from '@badeball/cypress-cucumber-preprocessor';
 
 const findActivationDates = () => {
@@ -45,4 +46,45 @@ Then('the email address input contains {string}', (emailAddress) => {
   cy.get('[data-cy=email-address-input]').then((elt) => {
     expect(elt.attr('value')).to.eql(emailAddress);
   });
+});
+
+Then(`I can export feedback and download it as a CSV file`, () => {
+  cy.intercept('POST', '/feedback?export=true', (req) => {
+    req.continue((res) => {
+      const contentDisposition = res.headers['content-disposition'];
+      const downloadedFile = contentDisposition.split('filename=')[1];
+      const downloadsFolder = Cypress.config('downloadsFolder');
+
+      cy.readFile(path.join(downloadsFolder, downloadedFile)).then(
+        (download) => {
+          const lines = download.split('\n');
+
+          // header line + 3 rows + 1 empty line
+          expect(lines.length).to.eql(5);
+          expect(lines[0]).to.eql(
+            'Received,From,"Phone number",Rating,Details,Page,Browser',
+          );
+
+          // check structure of rows after header
+          for (let i = 1; i < 4; i++) {
+            expect(lines[i]).to.match(
+              /"\d{2}\/\d{2}\/\d{4} \d{2}:\d{2}",.+,.+,.+,.+,\/.+,".+"/,
+            );
+          }
+        },
+      );
+    });
+  });
+
+  // work-around for cypress bug when downloading files from a link;
+  // see https://github.com/cypress-io/cypress/issues/7083#issuecomment-858489694
+  cy.document().then((doc) => {
+    doc.addEventListener('click', () => {
+      setTimeout(function () {
+        doc.location.reload();
+      }, 1000);
+    });
+  });
+
+  cy.contains('Export').click();
 });
